removed flywire connectome specific code
